fix auth state
