+++ src/pages/_app.tsx
@@ -10,8 +10,7 @@ import {AppProps} from 'next/app'
 import Head from 'next/head'
 import {FC} from 'react'
 
-import {UserContainer} from '@/utils/UserContainer'
-import {WebstromTokenContainer} from '@/utils/WebstromTokenContainer'
+import {AuthContainer} from '@/utils/AuthContainer'
 
 const MyApp: FC<AppProps> = ({Component, pageProps}) => {
   return (
@@ -20,12 +19,9 @@ const MyApp: FC<AppProps> = ({Component, pageProps}) => {
         <title>React App</title>
         <meta name="viewport" content="width=device-width, initial-scale=1" />
       </Head>
-      <WebstromTokenContainer.Provider>
-        {/* UserContainer needs to be a child of WebstromTokenContainer */}
-        <UserContainer.Provider>
-          <Component {...pageProps} />
-        </UserContainer.Provider>
-      </WebstromTokenContainer.Provider>
+      <AuthContainer.Provider>
+        <Component {...pageProps} />
+      </AuthContainer.Provider>
     </>
   )
 }
